core/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aniso8601==9.0.1
anyio==3.6.2
apturl==0.5.2
arabic-reshaper==3.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asn1crypto==1.5.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
Brlapi==0.7.0
cachetools==5.3.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.3
comm==0.1.2
command-not-found==0.3
cryptography==40.0.2
cssselect2==0.7.0
cupshelpers==1.0
datetime-truncate==1.1.1
dbus-python==1.2.16
debugpy==1.6.4
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.1.4
django-adaptors==0.2.5
django-admin-charts==1.2.4
django-allauth==0.51.0
django-autoslug==1.9.8
django-bower==5.2.0
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-defender==0.9.7
django-dotenv==1.4.2
django-elasticsearch-dsl==7.2.2
django-filter==22.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-memoize==2.3.1
django-multiselectfield==0.1.12
django-nvd3==0.9.7
django-phone-field==1.8.1
django-phonenumber-field==7.0.1
django-tinymce==3.5.0
django-tinymce4-lite==1.8.0
Django-Verify-Email==2.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
duplicity==0.8.12.0
elasticsearch==7.17.8
elasticsearch-dsl==7.4.0
entrypoints==0.3
executing==1.2.0
fasteners==0.14.1
fastjsonschema==2.16.2
filelock==3.9.0
flake8==6.0.0
frozenlist==1.3.3
future==0.18.2
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
html5lib==1.1
httplib2==0.22.0
idna==2.8
importlib-metadata==5.2.0
importlib-resources==5.10.1
inflection==0.5.1
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.3
jedi==0.18.2
Jinja2==3.1.2
jsmin==3.0.1
jsonschema==4.17.3
jupyter-client==7.4.8
jupyter-console==6.4.4
jupyter-core==5.1.0
jupyter-events==0.5.0
jupyter-server==2.0.1
jupyter-server-terminals==0.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.4
jwcrypto==1.4.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
monotonic==1.5
multidict==6.0.4
mysqlclient==2.1.1
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
netifaces==0.10.4
notebook==6.5.2
notebook-shim==0.2.2
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
openai==0.27.5
oscrypto==1.3.0
packaging==22.0
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
phonenumbers==8.13.2
pickleshare==0.7.5
Pillow==9.5.0
pip-autoremove==0.10.0
pkgutil-resolve-name==1.3.10
platformdirs==2.6.0
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==4.22.4
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.16.2
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
pycups==1.9.73
pyflakes==3.0.1
Pygments==2.13.0
PyGObject==3.36.0
pyHanko==0.17.2
pyhanko-certvalidator==0.22.0
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.9
pypdf==3.8.1
pypng==0.20220715.0
Pyrebase4==4.7.0
pyRFC3339==1.1
pyrsistent==0.19.2
python-apt==2.0.1+ubuntu0.20.4.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-decouple==3.8
python-dotenv==1.0.0
python-json-logger==2.0.4
python-jwt==4.0.0
python-nvd3==0.14.2
python-slugify==1.1.4
python3-openid==3.2.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==24.0.1
qrcode==7.4.2
qtconsole==5.4.0
QtPy==2.3.0
redis==4.5.5
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
requests-unixsocket==0.2.0
rsa==4.9
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.8.0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.3.2.post1
sqlparse==0.2.4
stack-data==0.6.2
svglib==1.5.1
systemd-python==234
terminado==0.17.1
text-unidecode==1.3
tinycss2==1.2.1
tornado==6.2
tqdm==4.65.0
traitlets==5.7.1
typing-extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.3.6
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.19.0
virtualenvwrapper-win==1.2.7
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.4
wrapt==1.15.0
xhtml2pdf==0.2.10
xkit==0.0.0
yarl==1.9.2
youtube-dl==2020.3.24
zipp==3.11.0
